rename and move pages one layer up
