+++ src/client/src/react/pages/Waiting.css
