+++ src/client/src/react/pages/Waiting.js
@@ -1,8 +1,8 @@
 import React from "react";
 import "./WaitForRound.css";
 
-import ConnectionStatus from "../../components/ConnectionStatus/ConnectionStatus";
-import Title from "../../components/Title";
+import ConnectionStatus from "../components/ConnectionStatus/ConnectionStatus";
+import Title from "../components/Title";
 
 function WaitForRound({ appTitle, isConnected }) {
   return (
